Simplify dbg_send() function by using a StringInfo.
At least some versions of MinGW vsnprintf() implementations don't correctly'
return the number of bytes required to hold a string after formatting, when
called with a NULL pointer. I'm not sure if this is a problem in any version
that we care about, but this makes the code cleaner anyway.