Remove unused variants of pldbg_get_target_info() function
It supported several different methods of specifying the target:
function/procedure name, or a trigger name, or a function/procedure oid.
But pgAdmin - the only consumer of this API - only ever used the oid-method,
so it's unlikely the other methods actually worked too well. If we ever
add support for looking up those other things in pgAdmin, I doubt we would
want to do it with this API. In fact, we probably should deprecate the
whole pldbg_get_target_info() function, in favor of looking up the
necessary information by querying the catalogs directly. That's what pgAdmin
does anyway to find information about objects in other contexts.