Remove plugin_profiler.
It's completely unrelated to the pldebugger plugin, so it should be a project
and module of its own. However, I don't think anyone is actually using it,
so we might as well just remove it.

If someone still cares about it, you can grab it from the git repository
history and resurrect it as a new project