Split out PL/pgSQL specific parts of plugin_debugger.c to separate .c file.
While no-one is working on e.g a plperl debugger, this is good to have in
the EnterpriseDB version that needs to support both PL/pgSQL and SPL. Makes
the code cleaner, anyway, IMHO.

This is just a pretty crude cut, moving functions from one file to another.
Much more is required to actually support multiple languages in the same
debugger session.