Don't get confused if client neglects to call pldbg_wait_for_breakpoint().
The API stipulates that after a call to pldbg_attach_to_port(), you must
call pldbg_wait_for_breakpoint() to read the status of the breakpoint that
the debuggee hit. That's fine, but if a client nevertheless doesn't call
pldbg_wait_for_breakpoint(), but goes straight to e.g pldbg_get_variables(),
pldbgapi module got all confused because it wrongly interpreted the
breakpoint string coming from the debuggee as being a response to the
pldbg_get_variables() command. The result was really strange error messages.

Make that more robust so that the breakpoint string is always read as
part of pldbg_attach_to_port(), and stashed in the debug session. It's no
longer required to call pldbg_wait_for_breakpoint() after
pldbg_attach_to_port(), although most clients will probably still want to do
that to read the current state.