Fix MSVC build.
The biggest change is the revelation that we cannot use MaxBackends on
Windows, because it's not marked as PGDLLIMPORT. We could probably change
that in PostgreSQL, but we want this to work with previous releases too. So
instead of allocating one slot for each backend, just have a bunch of slots
that are allocated on demand and released when not needed anymore.

This also change the MSVC project file to match the recent Makefile changes,
and misc sources fixes like using closesocket() instead of close() and
putting sockets into blocking mode (they seem to be non-blocking on Windows
by default - who would've thought..). These were correct before, but I
broke them in my refactorings.