Merge pldbgapi.so and plugin_debugger.so into one library.
It was a pain to build two .so files from one Makefile, and there was a
runtime dependency between the libraries, anyway, so it's much easier to
just put everything in one .so file.

While we're at it, there's no need to install plugin_debugger.so into the
$libdir/plugins directory. Just install it to $libdir like any other
extension library.