Add three EDB-only columns to the result set of pldbg_get_target_info().
The EDB-version of pldebugger has always had these fields, but the community
version has not. Now that we use the community version as is in EDB, we need
these fields.

Alternatively, we could adjust pgAdmin to fetch this information, but we're
so late in the release cycle that we want to avoid changes to pgAdmin. Maybe
in the future.

Unfortunately, now that we use an inline code block in the install script,
pldebugger no longer works against a pre-9.0 server. If you're using an
older server, you'll have to use an older version of pldebugger as well.